requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
aiohttp==3.9.5
    # via textual-dev
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   litestar
    #   openai
    #   watchfiles
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.4.1
    # via stack-data
attrs==23.2.0
    # via aiohttp
beanie==1.26.0
beautifulsoup4==4.12.3
blis==0.7.11
    # via thinc
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   beanie
    #   litestar
    #   rich-click
    #   textual-dev
    #   typer
    #   uvicorn
cloudpathlib==0.16.0
    # via weasel
confection==0.1.4
    # via
    #   thinc
    #   weasel
cryptography==42.0.7
    # via litestar
cymem==2.0.8
    # via
    #   preshed
    #   spacy
    #   thinc
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
distro==1.9.0
    # via openai
dnspython==2.6.1
    # via pymongo
docker==7.1.0
    # via
    #   testcontainers
    #   testcontainers-core
ecdsa==0.19.0
    # via python-jose
editorconfig==0.12.4
    # via jsbeautifier
executing==2.0.1
    # via stack-data
faker==25.2.0
    # via polyfactory
fast-query-parsers==1.0.3
    # via litestar
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   litestar
    #   openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
ipdb==0.13.13
ipython==8.24.0
    # via ipdb
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via
    #   litestar
    #   spacy
jsbeautifier==1.15.1
    # via litestar
langcodes==3.4.0
    # via spacy
language-data==1.2.0
    # via langcodes
lazy-model==0.2.0
    # via beanie
linkify-it-py==2.0.3
    # via markdown-it-py
litestar==2.8.3
loguru==0.7.2
marisa-trie==1.1.1
    # via language-data
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   rich
    #   textual
markupsafe==2.1.5
    # via jinja2
matplotlib-inline==0.1.7
    # via ipython
mdit-py-plugins==0.4.1
    # via markdown-it-py
mdurl==0.1.2
    # via markdown-it-py
motor==3.4.0
    # via beanie
msgpack==1.0.8
    # via textual-dev
msgspec==0.18.6
    # via litestar
multidict==6.0.5
    # via
    #   aiohttp
    #   litestar
    #   yarl
murmurhash==1.0.10
    # via
    #   preshed
    #   spacy
    #   thinc
numpy==1.26.4
    # via
    #   blis
    #   spacy
    #   thinc
openai==1.30.2
packaging==24.0
    # via
    #   pytest
    #   spacy
    #   thinc
    #   weasel
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pluggy==1.5.0
    # via pytest
polyfactory==2.16.0
    # via litestar
preshed==3.0.9
    # via
    #   spacy
    #   thinc
prompt-toolkit==3.0.43
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyasn1==0.6.0
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via
    #   beanie
    #   confection
    #   lazy-model
    #   openai
    #   pydantic-settings
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.18.2
    # via pydantic
pydantic-settings==2.2.1
pygments==2.18.0
    # via
    #   ipython
    #   rich
pymongo==4.7.2
    # via
    #   motor
    #   testcontainers-mongodb
pytest==8.2.1
    # via
    #   pytest-asyncio
    #   pytest-mock
pytest-asyncio==0.23.7
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
    # via faker
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-jose==3.3.0
    # via litestar
pyyaml==6.0.1
    # via
    #   litestar
    #   uvicorn
requests==2.32.2
    # via
    #   docker
    #   spacy
    #   weasel
rich==13.7.1
    # via
    #   litestar
    #   rich-click
    #   textual
rich-click==1.8.2
    # via litestar
rsa==4.9
    # via python-jose
ruff==0.4.5
setuptools==70.0.0
    # via
    #   marisa-trie
    #   spacy
    #   thinc
six==1.16.0
    # via
    #   asttokens
    #   ecdsa
    #   jsbeautifier
    #   python-dateutil
smart-open==6.4.0
    # via
    #   spacy
    #   weasel
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
spacy==3.7.4
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.4.8
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
stack-data==0.6.3
    # via ipython
testcontainers==4.4.1
testcontainers-core==0.0.1rc1
    # via testcontainers-mongodb
testcontainers-mongodb==0.0.1rc1
textual==0.63.2
    # via textual-dev
textual-dev==1.5.1
thinc==8.2.3
    # via spacy
toml==0.10.2
    # via beanie
tqdm==4.66.4
    # via
    #   openai
    #   spacy
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typer==0.9.4
    # via
    #   spacy
    #   weasel
typing-extensions==4.12.0
    # via
    #   ipython
    #   litestar
    #   openai
    #   polyfactory
    #   pydantic
    #   pydantic-core
    #   rich-click
    #   testcontainers
    #   textual
    #   textual-dev
    #   typer
uc-micro-py==1.0.3
    # via linkify-it-py
urllib3==2.2.1
    # via
    #   docker
    #   requests
    #   testcontainers
uvicorn==0.29.0
    # via litestar
uvloop==0.19.0
    # via
    #   litestar
    #   uvicorn
wasabi==1.1.2
    # via
    #   spacy
    #   thinc
    #   weasel
watchfiles==0.21.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
weasel==0.3.4
    # via spacy
websockets==12.0
    # via uvicorn
wrapt==1.16.0
    # via
    #   testcontainers
    #   testcontainers-core
yarl==1.9.4
    # via aiohttp
